ansible-iq/requirements.txt
aiojobs==1.4.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
fastapi==0.115.12
fastapi-jsonrpc==3.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
json-rpc==1.15.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mcp==1.8.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
platformdirs==4.3.8
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
sniffio==1.3.1
sse-starlette==2.3.4
starlette==0.46.2
textual==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uc-micro-py==1.0.3
urllib3==2.4.0
uvicorn==0.34.2
